Change linePrecision on raycaster
Default THREE.Raycaster linePrecision is 1, meaning that any InteractiveMarker with THREE.Line geometry will be selected when the mouse is within 1m of any point on any line in the marker, making interaction difficult or impossible. This changes the precision so that the marker is only highlighted when the mouse is actually hovering over a line in the marker.